+++ nemo/collections/asr/parts/submodules/multi_head_attention.py
@@ -37,7 +37,7 @@ import math
 import torch
 import torch.nn as nn
 
-from nemo.collections.common.parts.training_utils import avoid_float16_autocast_context
+from nemo.utils import avoid_float16_autocast_context
 
 __all__ = [
     'RelPositionMultiHeadAttention',
@@ -140,7 +140,8 @@ class MultiHeadAttention(nn.Module):
         """
         key, value, query = self.update_cache(key=key, value=value, query=query, cache=cache, cache_next=cache_next)
 
-        # temporary until we solve this more gracefully
+        if torch.is_autocast_enabled():
+            query, key, value = query.to(torch.float32), key.to(torch.float32), value.to(torch.float32)
         with avoid_float16_autocast_context():
             q, k, v = self.forward_qkv(query, key, value)
             scores = torch.matmul(q, k.transpose(-2, -1)) / self.s_d_k
